requirements.txt
tensorflow==2.20.0
flask==3.1.0
pillow==11.0.0
numpy==2.1.0
pandas==2.2.3
opencv-python==4.10.0.84
matplotlib==3.9.2
scikit-learn==1.5.2
requests==2.32.3